// src/app/interfaces/user-data.ts
export interface UserData {
    uid: string;
    name: string;
    lastName: string;
    idNumber: string;
    age: number;
    healthInsurance: string;
    email: string;
    password: string;
    password2: string;
    formFile: string;
    formFile2: string;
    role: string;    
    displayName: string;
}